requirements.txt
accelerate==0.29.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
boto3==1.34.80
botocore==1.34.80
bpemb==0.3.5
catalogue==2.0.10
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
configobj==5.0.8
configparser==6.0.1
conllu==4.5.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
emoji==2.11.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
fitz==0.0.1.dev2
flair==0.13.1
fonttools==4.50.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gdown==5.1.0
gensim==4.3.2
gliner==0.1.6
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
Janome==0.5.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.3
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kor==1.0.1
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.33
langchain-openai==0.1.0
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.29
llmsherpa==0.1.4
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
neo4j==5.18.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.14.2
orjson==3.9.15
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==3.1.0
pipreqs==0.5.0
platformdirs==4.2.0
pptree==3.1
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==5.26.1
prov==2.0.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytorch_revgrad==0.2.0
pytz==2024.1
pyvis==0.3.2
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.1
rdflib==7.0.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.7
rule-based-retrieval==0.1.3
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segtok==1.5.11
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
traits==6.3.2
transformer-smaller-training-vocab==0.4.0
transformers==4.38.2
triton==2.2.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unstructured==0.13.2
unstructured-client==0.18.0
urllib3==1.26.18
uvicorn==0.29.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websockets==11.0.3
whyhow==0.0.3
Wikipedia-API==0.6.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
